fix(hierarchy): 17 issues — critical cycle/orphan, high cap/search/drop, med/low guards
Critical:
- Ancestor cycle check before reparent (walk ancestor chain, abort if cycle)
- Delete re-roots children before Destroy (no more orphaned subtrees)

High:
- Remove silent 512-actor cap (use full actor_total)
- Dual drop path: drop_handled flag prevents same-frame double mutation
- Search no longer skips subtrees (row hidden but children always pushed)
- RemoveComponent<ParentComponent> guarded with HasComponent

Medium:
- 256→1024 child-per-parent tmp array
- Rename cancel only fires for the deleted actor's EntityID
- ForEach cap check removed (cap = actor_total)
- Tooltip Y clamped to 0 near panel top

Low:
- drop_handled applied to root drop zone row
- secure_strncpy return checked on rename commit; keeps rename active on fail
- Zero-actor no longer skips status bar
- Eye icon comment (needed for icon type, intentionally non-interactive)
- ContainsCI null haystack guard
- double timestamp for double-click (precision over long sessions)
- node_count replaces nc (naming clarity)

+++ Tetragrama/Panels/HierarchyPanel.h
@@ -43,7 +43,7 @@ namespace Tetragrama::Panels
         // so we track in the panel to avoid double-firing)
         uint32_t                      m_last_click_idx           = UINT32_MAX;
         uint32_t                      m_last_click_gen           = 0;
-        float                         m_last_click_time          = -1.f;
+        double                        m_last_click_time          = -1.0;
 
         bool IsCollapsed(ZEngine::ECS::EntityID id) const;
         void ToggleCollapsed(ZEngine::ECS::EntityID id);
